+++ detox-copilot/src/errors/CopilotError.ts
