+++ detox-copilot/src/types/index.ts
